pyproject.toml
[project]
name = "Healthclaw"
version = "0.2.0"
description = "A private AI wellbeing companion that runs locally with Gemma via Ollama. Onboard your whole family — each member gets their own isolated bot."
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Healthclaw Contributors"}
]
keywords = ["ai", "wellbeing", "companion", "health", "privacy", "local-ai", "gemma", "ollama"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
urls = { Homepage = "https://github.com/vlbandara/healthclaw", Repository = "https://github.com/vlbandara/healthclaw", Issues = "https://github.com/vlbandara/healthclaw/issues" }

dependencies = [
    "typer>=0.20.0,<1.0.0",
    "anthropic>=0.45.0,<1.0.0",
    "pydantic>=2.12.0,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "websockets>=16.0,<17.0",
    "websocket-client>=1.9.0,<2.0.0",
    "httpx>=0.28.0,<1.0.0",
    "ddgs>=9.5.5,<10.0.0",
    "oauth-cli-kit>=0.1.3,<1.0.0",
    "loguru>=0.7.3,<1.0.0",
    "readability-lxml>=0.8.4,<1.0.0",
    "rich>=14.0.0,<15.0.0",
    "croniter>=6.0.0,<7.0.0",
    "dingtalk-stream>=0.24.0,<1.0.0",
    "python-telegram-bot[socks]>=22.6,<23.0",
    "lark-oapi>=1.5.0,<2.0.0",
    "socksio>=1.0.0,<2.0.0",
    "python-socketio>=5.16.0,<6.0.0",
    "msgpack>=1.1.0,<2.0.0",
    "slack-sdk>=3.39.0,<4.0.0",
    "slackify-markdown>=0.2.0,<1.0.0",
    "qq-botpy>=1.2.0,<2.0.0",
    "python-socks[asyncio]>=2.8.0,<3.0.0",
    "prompt-toolkit>=3.0.50,<4.0.0",
    "questionary>=2.0.0,<3.0.0",
    "mcp>=1.26.0,<2.0.0",
    "json-repair>=0.57.0,<1.0.0",
    "chardet>=3.0.2,<6.0.0",
    "openai>=2.8.0",
    "tiktoken>=0.12.0,<1.0.0",
    "jinja2>=3.1.0,<4.0.0",
    "dulwich>=0.22.0,<1.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn>=0.34.0,<1.0.0",
    "cryptography>=45.0.0,<46.0.0",
    "aiosqlite>=0.20.0,<1.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "docker>=7.0.0,<8.0.0",
    "arq>=0.26.0,<1.0.0",
    "qrcode[pil]>=8.0",
    "alembic>=1.16.0,<2.0.0",
    "SQLAlchemy>=2.0.0,<3.0.0",
    "prometheus-client>=0.22.0,<1.0.0",
    "redis[hiredis]>=4.2.0,<6.0.0",
    "slowapi>=0.1.9,<1.0.0",
    "langfuse>=3.0.0,<4.0.0",
    "mem0ai>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
api = [
    "aiohttp>=3.9.0,<4.0.0",
]
wecom = [
    "wecom-aibot-sdk-python>=0.1.5",
]
weixin = [
    "pycryptodome>=3.20.0",
]

matrix = [
    "matrix-nio[e2e]>=0.25.2",
    "mistune>=3.0.0,<4.0.0",
    "nh3>=0.2.17,<1.0.0",
]
discord = [
    "discord.py>=2.5.2,<3.0.0",
]
langsmith = [
    "langsmith>=0.1.0",
]
dev = [
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "aiohttp>=3.9.0,<4.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
nanobot = "nanobot.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = [
    "nanobot/**/*.py",
    "nanobot/templates/**/*.md",
    "nanobot/skills/**/*.md",
    "nanobot/skills/**/*.sh",
    "nanobot/health/templates/**/*.html",
    "nanobot/health/static/**/*",
]

[tool.hatch.build.targets.wheel]
packages = ["nanobot"]

[tool.hatch.build.targets.wheel.sources]
"nanobot" = "nanobot"

[tool.hatch.build.targets.wheel.force-include]
"bridge" = "nanobot/bridge"

[tool.hatch.build.targets.sdist]
include = [
    "nanobot/",
    "bridge/",
    "README.md",
    "LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["nanobot"]
omit = ["tests/*", "**/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
